added progress component
